requirement.txt
bcrypt                    4.3.0       
flasgger                  0.9.7.1     
Flask                     3.1.1       
Flask-PyMongo             3.0.1       
Jinja2                    3.1.6       
jsonschema                4.25.0      
jsonschema-specifications 2025.4.1    
MarkupSafe                3.0.2       
mistune                   3.1.3       
packaging                 25.0        
pip                       25.2        
pymongo                   4.14.0      
python-dotenv             1.1.1       
PyYAML                    6.0.2
referencing               0.36.2      
rpds-py                   0.26.0      
setuptools                75.1.0      
six                       1.17.0      
typing_extensions         4.14.1      
Werkzeug                  3.1.3       
wheel                     0.44.0